requirements.txt
flask
python-dotenv
openai
requests
sentence-transformers
nltk
sqlalchemy
chromadb
pydantic